back/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
sqlalchemy>=2.0.21
pydantic>=2.4.2
pydantic[email]>=2.4.2
bcrypt>=4.0.0
pyjwt>=2.8.0
python-multipart>=0.0.6
websockets>=11.0.3
